refactor(extension): extract shared open-perapera-db helper from IndexedDbSessionStore
domain Repository adapter (DD-260〜263) 群と IndexedDbSessionStore (DD-106) が
同一 IndexedDB を共有するため、schema / store 定数 / open helper を
open-perapera-db.ts に切り出し。TranscriptStream 組立ロジックも
transcript-stream-assembler.ts に抽出。public API は無変更、既存 27 tests green。
