+++ packages/extension/src/infrastructure/storage/indexed-db-session-store.ts
@@ -1,92 +1,21 @@
-import { openDB, type DBSchema, type IDBPDatabase } from 'idb';
-import { ResultAsync, err, errAsync, ok, okAsync, type Result } from 'neverthrow';
-import {
-  invariantViolationError,
-  notFoundError,
-  type DomainError,
-} from '../../domain/shared/errors';
-import { type SessionIdentifier } from '../../domain/session/session-identifier';
-import { type TranscriptSegment } from '../../domain/transcript/transcript-segment';
-import { type TranscriptStream } from '../../domain/transcript/transcript-stream';
-import { type TranslationSegment } from '../../domain/transcript/translation-segment';
+import { ResultAsync, errAsync, okAsync } from 'neverthrow';
+import { notFoundError, type DomainError } from '../../domain/shared/errors';
 import { type ExportBundle, type SessionStore } from '../../application/ports/session-store';
+import {
+  INDEXED_DB_NAME,
+  SESSIONS_STORE,
+  TRANSCRIPT_STORE,
+  TRANSLATION_STORE,
+  createPeraperaDbHandle,
+  toPersistenceError,
+} from './open-perapera-db';
 import {
   sessionFromRecord,
   sessionToRecord,
-  transcriptSegmentFromRecord,
   transcriptSegmentToRecord,
-  translationSegmentFromRecord,
   translationSegmentToRecord,
-  type SessionRow,
-  type TranscriptSegmentRow,
-  type TranslationSegmentRow,
 } from './records';
-
-export const INDEXED_DB_NAME = 'perapera';
-export const INDEXED_DB_VERSION = 1;
-
-const SESSIONS_STORE = 'sessions';
-const TRANSCRIPT_STORE = 'transcript_segments';
-const TRANSLATION_STORE = 'translation_segments';
-const EXPORT_STORE = 'export_records';
-
-interface PeraperaSchema extends DBSchema {
-  sessions: {
-    key: string;
-    value: SessionRow;
-  };
-  transcript_segments: {
-    key: string;
-    value: TranscriptSegmentRow;
-    indexes: { 'by-sessionId': string };
-  };
-  translation_segments: {
-    key: string;
-    value: TranslationSegmentRow;
-    indexes: { 'by-sessionId': string };
-  };
-  export_records: {
-    key: string;
-    value: {
-      exportId: string;
-      sessionId: string;
-      format: 'txt' | 'json';
-      includeOriginal: boolean;
-      includeTranslation: boolean;
-      createdAt: string;
-    };
-    indexes: { 'by-sessionId': string };
-  };
-}
-
-const openPeraperaDb = (databaseName: string): Promise<IDBPDatabase<PeraperaSchema>> =>
-  openDB<PeraperaSchema>(databaseName, INDEXED_DB_VERSION, {
-    upgrade(db) {
-      if (!db.objectStoreNames.contains(SESSIONS_STORE)) {
-        db.createObjectStore(SESSIONS_STORE, { keyPath: 'sessionId' });
-      }
-      if (!db.objectStoreNames.contains(TRANSCRIPT_STORE)) {
-        const store = db.createObjectStore(TRANSCRIPT_STORE, { keyPath: 'segmentId' });
-        store.createIndex('by-sessionId', 'sessionId');
-      }
-      if (!db.objectStoreNames.contains(TRANSLATION_STORE)) {
-        const store = db.createObjectStore(TRANSLATION_STORE, { keyPath: 'translationId' });
-        store.createIndex('by-sessionId', 'sessionId');
-      }
-      if (!db.objectStoreNames.contains(EXPORT_STORE)) {
-        const store = db.createObjectStore(EXPORT_STORE, { keyPath: 'exportId' });
-        store.createIndex('by-sessionId', 'sessionId');
-      }
-    },
-  });
-
-const toPersistenceError =
-  (scope: string) =>
-  (cause: unknown): DomainError =>
-    invariantViolationError({
-      invariant: 'session-persistence',
-      details: `${scope}: ${cause instanceof Error ? cause.message : 'unknown error'}`,
-    });
+import { assembleTranscriptStream } from './transcript-stream-assembler';
 
 /**
  * IndexedDB session store の拡張 interface。
@@ -113,6 +42,10 @@ export type CloseableSessionStore = SessionStore & {
  * 非同期 append-only 設計 (CLAUDE.md §データ保存方針)。ホットパスから
  * fire-and-forget で呼ばれる想定で、書き込み失敗は UseCase 層で WARN ログ
  * に留める。
+ *
+ * DB 共通定義 (schema / 定数 / open helper) は `open-perapera-db.ts` に集約し、
+ * domain Repository adapter 群 (`indexed-db-source-session-repository.ts` など)
+ * と共有する。
  */
 export type IndexedDbSessionStoreOptions = Readonly<{
   /** Override for tests. Production should omit this to use the default. */
@@ -123,42 +56,13 @@ export const createIndexedDbSessionStore = (
   options: IndexedDbSessionStoreOptions = {},
 ): CloseableSessionStore => {
   const databaseName = options.databaseName ?? INDEXED_DB_NAME;
-  let dbPromise: Promise<IDBPDatabase<PeraperaSchema>> | null = null;
-  const db = (): Promise<IDBPDatabase<PeraperaSchema>> => {
-    dbPromise ??= openPeraperaDb(databaseName);
-    return dbPromise;
-  };
-
-  const buildStream = (
-    sessionIdentifier: SessionIdentifier,
-    transcripts: readonly TranscriptSegmentRow[],
-    translations: readonly TranslationSegmentRow[],
-  ): Result<TranscriptStream, DomainError> => {
-    const segments = new Map<string, TranscriptSegment>();
-    for (const row of transcripts) {
-      const result = transcriptSegmentFromRecord(row);
-      if (result.isErr()) return err(result.error);
-      segments.set(result.value.segmentIdentifier, result.value);
-    }
-    const translationMap = new Map<string, TranslationSegment>();
-    for (const row of translations) {
-      const result = translationSegmentFromRecord(row);
-      if (result.isErr()) return err(result.error);
-      translationMap.set(result.value.segmentIdentifier, result.value);
-    }
-    const stream: TranscriptStream = {
-      sessionIdentifier,
-      segments,
-      translations: translationMap,
-    };
-    return ok(stream);
-  };
+  const handle = createPeraperaDbHandle(databaseName);
 
   return {
     saveSession: (session) =>
       ResultAsync.fromPromise(
         (async () => {
-          const connection = await db();
+          const connection = await handle.get();
           await connection.put(SESSIONS_STORE, sessionToRecord(session));
         })(),
         toPersistenceError('saveSession'),
@@ -167,7 +71,7 @@ export const createIndexedDbSessionStore = (
     appendTranscript: (sessionIdentifier, segment) =>
       ResultAsync.fromPromise(
         (async () => {
-          const connection = await db();
+          const connection = await handle.get();
           await connection.put(
             TRANSCRIPT_STORE,
             transcriptSegmentToRecord(sessionIdentifier, segment),
@@ -179,7 +83,7 @@ export const createIndexedDbSessionStore = (
     appendTranslation: (sessionIdentifier, translation) =>
       ResultAsync.fromPromise(
         (async () => {
-          const connection = await db();
+          const connection = await handle.get();
           await connection.put(
             TRANSLATION_STORE,
             translationSegmentToRecord(sessionIdentifier, translation),
@@ -191,7 +95,7 @@ export const createIndexedDbSessionStore = (
     loadExportBundle: (sessionIdentifier) =>
       ResultAsync.fromPromise(
         (async () => {
-          const connection = await db();
+          const connection = await handle.get();
           const sessionRow = await connection.get(SESSIONS_STORE, sessionIdentifier);
           if (sessionRow === undefined) return null;
           const transcripts = await connection.getAllFromIndex(
@@ -217,7 +121,11 @@ export const createIndexedDbSessionStore = (
         if (sessionResult.isErr()) {
           return errAsync<ExportBundle, DomainError>(sessionResult.error);
         }
-        const streamResult = buildStream(sessionIdentifier, raw.transcripts, raw.translations);
+        const streamResult = assembleTranscriptStream(
+          sessionIdentifier,
+          raw.transcripts,
+          raw.translations,
+        );
         if (streamResult.isErr()) {
           return errAsync<ExportBundle, DomainError>(streamResult.error);
         }
@@ -227,12 +135,8 @@ export const createIndexedDbSessionStore = (
         });
       }),
 
-    close: async () => {
-      if (dbPromise !== null) {
-        const connection = await dbPromise;
-        connection.close();
-        dbPromise = null;
-      }
-    },
+    close: handle.close,
   };
 };
+
+export { INDEXED_DB_NAME, INDEXED_DB_VERSION } from './open-perapera-db';
